urban-server/urban_design_backend/requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dj-rest-auth==6.0.0
Django==5.0.4
django-allauth==0.62.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
fonttools==4.51.0
google-auth==2.31.0
gunicorn==22.0.0
idna==3.7
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pillow==10.3.0
platformdirs==4.2.1
pooch==1.8.1
psycopg==3.1.19
psycopg-binary==3.1.19
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
scooby==0.10.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
